+++ tests/asyncio/test_connection.py
@@ -3,7 +3,7 @@ from sqlalchemy import select
 from sqlalchemy.exc import StatementError
 from sqlalchemy.schema import CreateTable
 
-from sqlalchemy_aio._base import AsyncTransaction
+from sqlalchemy_aio.base import AsyncTransaction
 
 
 @pytest.mark.asyncio